Version 2.6
Notifikasi

+++ zerolevel2/controllers/pegawai/Dokumen.php
@@ -6,61 +6,42 @@ class Dokumen extends CI_Controller {
 	public function __construct() {
 		parent::__construct();
 
-		//* Check if current-user is authorized user *//
-		if (!isset($this->session->userdata['logged_in_portal']['auth'])) {
+		//* Check if current-user is pegawai *//
+		if (!isset($this->session->userdata['logged_in_portal']['pgw'])) {
 			if (!isset($this->session->userdata['logged_in_portal'])) {
 				redirect(site_url('login')."?redirect=".uri_string());
 			} else {
 				show_error('Access denied!');
 			}
-		}
+		}	
 
 		//* Load model, library, helper, etc *//
-		$this->load->model(array('Tabel_dokumen','Tabel_dosen','Tabel_refDocgroup','Tabel_xNotifikasi'));
+		$this->load->model(array('Tabel_dokumen','Tabel_refDocgroup','Tabel_xNotifikasi'));
 
 		//* Initialize class variables. current-user identity. To be used throughout this class *//
 		$this->user = array(
-			'nama' 		=> $this->session->userdata['logged_in_portal']['auth']['nama'],
-			'userid'	=> $this->session->userdata['logged_in_portal']['auth']['userid'],
-			'kodeunit'	=> $this->session->userdata['logged_in_portal']['auth']['kodeUnit'],
-			'namaunit'	=> $this->session->userdata['logged_in_portal']['auth']['namaUnit'],
-			'grup' 		=> $this->session->userdata['logged_in_portal']['auth']['grup'],
-			'kodegrup'  => $this->session->userdata['logged_in_portal']['auth']['kodeGrup'],
-		);		
-		
+			'nama' 		=> $this->session->userdata['logged_in_portal']['pgw']['nama'],
+			'id'		=> $this->session->userdata['logged_in_portal']['pgw']['userid'],
+		);	
 	}
-
+	
 	public function index() {
 
 		//* Initialize general variables for pageview properties *//
-		$data['pageTitle'] 	= "Repository";
-		$data['subtitle'] 	= $this->user['namaunit'];
-		$data['body_page'] 	= "body/dokumen/list_with_add_multiple";
-
+		$data['pageTitle'] 	= "My Documents";
+		$data['subtitle'] 	= "Daftar Dokumen";
+		$data['body_page'] 	= "body/dokumen/list_with_add_multiple_pegawai";
+		
 		//* Declare variables array $data to be passing to view *//
-		$data['dokumen'] 	= $this->Tabel_dokumen->get("ownerId REGEXP '[a-zA-Z]'",'dokumenTahun DESC, dokumenDocgroupId ASC');
+		$data['dokumen'] 	= $this->Tabel_dokumen->user_get(array('aso_dokumen.userId'=> $this->user['id']),'dokumenTahun DESC, dokumenDocgroupId ASC');
 		$data['docgroup'] 	= $this->Tabel_refDocgroup->get();
-		$data['ownerId']	= $this->user['kodegrup'];
-		$data['fileSpec']	= "Filetype= pdf jpg jpeg xls xlsx mde doc docx; Max Size=20 Mb.";
-
-		switch ($this->user['kodegrup']) {
-			case "dekan"		: $data['nodeId'] = '0'; break;
-			case "wd1"			: $data['nodeId'] = '1'; break;
-			case "wd2"			: $data['nodeId'] = '2'; break;
-			case "wd3"			: $data['nodeId'] = '3'; break;
-			case "jurusan45"	: $data['nodeId'] = '5'; break;
-			case "jurusan42"	: $data['nodeId'] = '6'; break;
-			case "jurusan43"	: $data['nodeId'] = '7'; break;
-			case "jurusan44"	: $data['nodeId'] = '8'; break;
-			case "prodi14"		: $data['nodeId'] = '10'; break;
-			case "prodi94"		: $data['nodeId'] = '11'; break;
-			case "prodi15"		: $data['nodeId'] = '12'; break;
-			case "prodi16"		: $data['nodeId'] = '13'; break;
-			case "prodi12"		: $data['nodeId'] = '14'; break;
-			case "prodi77"		: $data['nodeId'] = '15'; break;
-			case "prodi13"		: $data['nodeId'] = '16'; break;
-			default 			: $data['nodeId'] = '17'; break;
-		}
+		$data['ownerId']	= $this->user['id'];
+		$data['loadMe']		= array(
+									'tipe' => 'peg',
+									'nama' => $this->user['nama'],
+									'id' => $this->user['id']
+									);
+		$data['fileSpec']	= "Filetype = pdf jpg jpeg; Max Size = 5 Mb.";
 		
 		//* formatting the data to be view properly at the pageview *//
 		foreach ($data['dokumen'] as &$val) {
@@ -69,60 +50,13 @@ class Dokumen extends CI_Controller {
 		
 		$this->load->view(THEME,$data);
 
-	}	
-
-	public function dosen($nip=FALSE) {
-
-		if ($nip === FALSE) {
-			
-			//* Switch data dosen between: prodi, jurusan and fakultas *//
-			switch ($this->user['grup']) {
-
-				//* Get data dosen from local database and return to $dosen *//
-				case "prodi"	: $data['dosen'] = $this->Tabel_dosen->get(array('kodeProdi' => $this->user['kodeunit'])); break;
-				case "jurusan"	: $data['dosen'] = $this->Tabel_dosen->get(array('kodeJurusan' => $this->user['kodeunit'])); break;
-				default 		: $data['dosen'] = $this->Tabel_dosen->get(); break;
-
-			}
-
-			//* formatting the data to be view properly at the pageview *//
-			foreach ($data['dosen'] as $key => $value) {
-				$data['dosen'][$key]['jurusan'] = ucwords(strtolower($value['jurusan']));
-				$data['dosen'][$key]['prodi'] = ucwords(strtolower($value['prodi']));
-			}
-
-			//* Initialize general variables for pageview properties *//
-			$data['pageTitle'] 	= "Dokumen Dosen ". $this->user['namaunit'];
-			$data['subtitle'] 	= "Daftar Dosen";
-			$data['body_page'] 	= "body/dokumen/list_dosen";
-
-		} else {
-
-			$namaDosen = $this->Tabel_dosen->detail(array('nip' => $nip))['nama'];
-
-			//* Initialize general variables for pageview properties *//
-			$data['pageTitle'] 	= "Dokumen ". $this->user['namaunit'];
-			$data['subtitle'] 	= "Daftar Dokumen ". $namaDosen;
-			$data['body_page'] 	= "body/dokumen/list_without_add";
-
-			//* Declare variables array $data to be passing to view *//
-			$data['dokumen'] 	= $this->Tabel_dokumen->user_get(array('aso_dokumen.userId'=> $nip),'dokumenTahun DESC, dokumenDocgroupId ASC');
-			$data['docgroup'] 	= $this->Tabel_refDocgroup->get();
-
-			//* formatting the data to be view properly at the pageview *//
-			foreach ($data['dokumen'] as &$val) {
-				$val['dokumenFile'] = URL_DOKUMEN.$val['dokumenFile'];
-			}
-		}
-
-		$this->load->view(THEME,$data);
 	}
 
 	public function tambah() {
 
 		//* Set rules for form validation. Form validation use CI Library *//
 		$this->form_validation->set_rules('nama', 'Nama Dokumen', 'trim|required');
-		$this->form_validation->set_rules('dsndoc', 'Tag Dosen', 'required');
+		$this->form_validation->set_rules('pegdoc', 'Tag Pegawai', 'required');
 		if (empty($_FILES['dokumen']['name'])) {
 			$this->form_validation->set_rules('dokumen', 'Upload Dokumen', 'required');
 		}		
@@ -136,14 +70,14 @@ class Dokumen extends CI_Controller {
 			$database['dokumenNomor'] 		= $this->input->post('nomor');
 			$database['dokumenTahun']		= $this->input->post('tahun');
 			$database['dokumenDocgroupId'] 	= $this->input->post('jenis');
-			$database['ownerId'] 			= $this->user['kodegrup'];
+			$database['ownerId'] 			= $this->user['id'];
 
 			//* Check if dokumen has submit *//
 			if(!empty($_FILES['dokumen']['name'])) {
 
 				//* Set the configuration for upload library. Set the file name. with format idUser-currentDate.jpg *//
-				$this->config->config['dokumen_admin']['file_name'] = $this->user['kodegrup']."-".date('Ymd');
-				$this->load->library('upload', $this->config->item('dokumen_admin'));
+				$this->config->config['dokumen']['file_name'] = $this->user['id']."-".date('Ymd');
+				$this->load->library('upload', $this->config->item('dokumen'));
 
 				//* Check if the dokumen has not been upload *//
 				if(!$this->upload->do_upload('dokumen')) {
@@ -184,7 +118,7 @@ class Dokumen extends CI_Controller {
 					
 					//* Input the usertag into database *//
 					if ($this->Tabel_dokumen->link_user($multi_user)) {
-
+					
 						$notif['tipe'] = "dokumen";
 						$notif['isiNotif'] = "Dokumen baru, anda ditandai!";
 
@@ -229,7 +163,7 @@ class Dokumen extends CI_Controller {
 			$this->session->set_flashdata('message', validation_errors('Gagal disimpan. Form tidak lengkap! '));
 		}
 
-		redirect(site_url('dokumen'));
+		redirect(site_url('dosen/dokumen'));	
 
 	}
 
@@ -237,7 +171,7 @@ class Dokumen extends CI_Controller {
 
 		//* Set rules for form validation. Form validation use CI Library *//
 		$this->form_validation->set_rules('nama', 'Nama Dokumen', 'trim|required');
-		$this->form_validation->set_rules('dsndoc', 'Tag Dosen', 'required');
+		$this->form_validation->set_rules('pegdoc', 'Tag Pegawai', 'required');
 		
 		//* Check if form valid *//
 		if ($this->form_validation->run() == TRUE) {
@@ -249,14 +183,14 @@ class Dokumen extends CI_Controller {
 			$database['dokumenNomor'] 		= $this->input->post('nomor');
 			$database['dokumenTahun']		= $this->input->post('tahun');
 			$database['dokumenDocgroupId'] 	= $this->input->post('jenis');
-			$database['userUpdate']			= $this->user['nama'];
+			$database['userUpdate']			= $this->user['id'];
 
 			//* If new dokumen is submitted, perform below procedure. If not, do nothing *//
 			if(!empty($_FILES['dokumen']['name'])) {
 
 				//* Set the configuration for upload library. Set the file name. with format idUser-currentDate.jpg *//
-				$this->config->config['dokumen_admin']['file_name'] = $this->user['kodegrup']."-".date('Ymd');
-				$this->load->library('upload', $this->config->item('dokumen_admin'));
+				$this->config->config['dokumen']['file_name'] = $this->user['id']."-".date('Ymd');
+				$this->load->library('upload', $this->config->item('dokumen'));
 
 				//* Check if the dokumen has not been upload *//
 				if(!$this->upload->do_upload('dokumen')) {
@@ -292,6 +226,10 @@ class Dokumen extends CI_Controller {
 					$usermhs = explode(",", $this->input->post('mhsdoc'));
 					foreach ($usermhs as $key) array_push($user, $key);
 				}
+				if ($this->input->post('pegdoc')) {
+					$userpeg = explode(",", $this->input->post('pegdoc'));
+					foreach ($userpeg as $key) array_push($usertag, $key);
+				}
 
 				//* Declare array multi_user and formatting data, which will inserted into database*//
 				foreach ($user as $val) {
@@ -325,7 +263,7 @@ class Dokumen extends CI_Controller {
 			$this->session->set_flashdata('message', validation_errors('Form tidak lengkap! '));
 		}
 
-		redirect(site_url('dokumen'));
+		redirect(site_url('dosen/dokumen'));
 	
 	}
 
@@ -336,23 +274,41 @@ class Dokumen extends CI_Controller {
 		//* Get detail dokumen with id = $id_dokumen *//
 		$file = $this->Tabel_dokumen->detail(array('ft_dokumen.idDokumen' => $id_dokumen));
 
-		//* Delete dokumen file*//
-		if (file_exists(FCPATH.DIR_DOKUMEN.$file['dokumenFile'])) unlink(FCPATH.DIR_DOKUMEN.$file['dokumenFile']);
+		//* Check if owner files is this user *//
+		if ($file['ownerId'] == $this->user['id']) {
 
-		//* Delete entry in database *//
-		if ($this->Tabel_dokumen->delete($id_dokumen)) {
-				
-			$this->session->set_flashdata('type', 'success');
-			$this->session->set_flashdata('message', 'Berhasil dihapus!');
+			//* Owner files is this user: delete dokumen file*//
+			if (file_exists(FCPATH.DIR_DOKUMEN.$file['dokumenFile'])) unlink(FCPATH.DIR_DOKUMEN.$file['dokumenFile']);
+
+			//* Delete entry in database *//
+			if ($this->Tabel_dokumen->delete($id_dokumen)) {
+					
+				$this->session->set_flashdata('type', 'success');
+				$this->session->set_flashdata('message', 'Berhasil dihapus!');
+
+			} else {
+			
+				$this->session->set_flashdata('type', 'danger');
+				$this->session->set_flashdata('message', 'Gagal dihapus!');
+				$this->output->set_status_header('500');
+			}
 
 		} else {
-		
-			$this->session->set_flashdata('type', 'danger');
-			$this->session->set_flashdata('message', 'Gagal dihapus!');
-			$this->output->set_status_header('500');
+			//* Owner files is not this user. Delete entry in database just this user*//
+			if ($this->Tabel_dokumen->unlink_user($id_dokumen,$this->user['id'])) {
+					
+				$this->session->set_flashdata('type', 'success');
+				$this->session->set_flashdata('message', 'Berhasil dihapus!');
+
+			} else {
+			
+				$this->session->set_flashdata('type', 'danger');
+				$this->session->set_flashdata('message', 'Gagal dihapus!');
+				$this->output->set_status_header('500');
+			}
 		}
 
-	}	
+	}
 
 	public function detail($id) {
